requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.7.2
    # via
    #   django
    #   django-cors-headers
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
coreapi==2.3.3
    # via -r requirements.in
coreschema==0.0.4
    # via coreapi
django==5.0.3
    # via
    #   -r requirements.in
    #   django-ckeditor-5
    #   django-cors-headers
    #   django-extensions
    #   django-geojson
    #   django-leaflet
    #   django-polymorphic
    #   djangorestframework
    #   drf-yasg
    #   jsonfield
django-ckeditor-5==0.2.11
    # via -r requirements.in
django-cors-headers==4.3.1
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-geojson==4.0.0
    # via -r requirements.in
django-leaflet==0.29.0
    # via -r requirements.in
django-polymorphic==3.1.0
    # via -r requirements.in
django-sortedm2m==3.1.1
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   drf-extensions
    #   drf-yasg
drf-extensions==0.7.1
    # via -r requirements.in
drf-yasg==1.21.7
    # via -r requirements.in
idna==3.7
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.1.3
    # via coreschema
jsonfield==3.1.0
    # via -r requirements.in
markdown==3.5.2
    # via -r requirements.in
markupsafe==2.1.5
    # via jinja2
packaging==23.2
    # via drf-yasg
pillow==10.3.0
    # via
    #   -r requirements.in
    #   django-ckeditor-5
psycopg2==2.9.9
    # via -r requirements.in
pytz==2024.1
    # via
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0.1
    # via drf-yasg
raven==6.10.0
    # via -r requirements.in
requests==2.31.0
    # via coreapi
sqlparse==0.5.0
    # via django
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==2.2.1
    # via requests
